Separating ATS core from ATS extensions (#11)
Update test/HelloATS/ats_hello to work on blueos
as well as on toss.
Add HelloGPU test.

Need to update the GPU test to do a bit more,
including showing the GPU mapping to MPI ranks.

Remove setup.py

Use setup-py2.py to install.  See instructions in Discussion.
I will update the README or other file as well stating how to
use setup-py2.py.

This setup.py was one David had been working for a new installation
method, but that is not ready yet.  At this time, we have a backup
which may be used in another repo, but it may be best just to start
over.