move files to root dir (#34)
* WIP: move files to root dir

* Slurm version aware and other fixes. (#33)

Update slurm module so it knows
what slurm version is being used. This
is needed due to differences in how
slurm behaves with the update by SchedMD.
The new slurm is on one machine for testing,
other machines have the older slurm.

Fix print statement.  A '%' character
had been lost in the conversion of print
statements to be python 3 compatible.

Remove use of --share when construction
srun line.  This option is deprecated
and is not supported in current slurm.

Added --unbuffered option at the request
of ATS user.  If specified, will add
-unbuffered to the srun line.  The user
wants to see if this helps them better
view the output from a run.  This is off
by default.

Change miscellaneous use of "-v" when
constructiong the srun line to use
--comment="xxx".  These options can
not be blank, as that breaks the run
under python.  They have to be something.
I was using the -v, which turns on
a more verbose srun run.  But the --comment
is a better no-op in these situations.

Placeholder to better explore --exclusve=user
option. I have not implemented this, as in
my testing it results in some interconnect
errors.  This was suggested by ans ATS user
as well.

* WIP: move files to root dir

Co-authored-by: Shawn Dawson <dawson6@llnl.gov>
Co-authored-by: David Bloss <davidbloss@Davids-Air.lan>