Various updates and bug fixes.
Default ruby machine type to slurm56

Add Krike testing to Python 3 branch

For slurm:
Account for slurm version such as 21.08.8-2

Added --useMinNodes for toss (slurm)

If --useMinNodes specified, then within an allocation
specify

    srun_nodes="--nodes=%i-%i" % (minNodes, minNodes)

when starting the job, where minNodes is the minimum
number of nodes needed for the requested number of MPI
processes.   This is experimental at this point,
and may lead to hangs or lower throughput.

For blueos:
Add smpi options to ats command line

Either one of these may be used to disable the --smpiargs="-gpu" option.

--smpi_off
--smpi_show