Improving Flux scheduler.
Add --exclusive to the flux
mini alloc command line.
May not be needed, but lets
be explicit.

Do not default to -N 1 when
in flux run line.  This was inhibiting
throughput by asking for exclusive node
access as the default for each test job.

Default the native scheduler to flux
on certain machines.

Update the installed version to 7.0.111
for alpha testing.

Update test examples