Update HelloATS READ.ME
Fix detection of 'am I in a Flux allocation?'

It used to be that one could look for this env var

   if "FLUX_JOB_ID" in os.environ

But that env var is now apparently not there with the flux
installed on rzwhippet.  So look for alternatives that
are not present on the login node, but are withn a
flux allocation

   if  "FLUX_JOB_ID" in os.environ or "FLUX_CONNECTOR_PATH" in os.environ or "FLUX_TERMINUS_SESSION" in os.environ

Update default number of flux nodes to 2.