Updates to fluxSchedule.py
Remove hard coded flux run options
 " -o cpu-affinity=per-task -o mpibind=off"
as options such as these may vary by project.

Added flux_run_args processing.  This is a string
that may be specified on the ats command line
for use with flux, and they will be appended
to the flux run line.  This will allow, for instance,
a project to add the option
"-o cpu-affinity=per-task -o mpibind=off"
(or any other flux run option) as works best for
their project.