Add -N option to the flux run line.
Honor the -nn option if specified on a per test
case, or on the command line.  If note specified,
then determine the proper number to be used
to set the -N, based on the number of cores needed
to run the jobs and the number of cores per node.

Without the -N option, flux was spewing mpi ranks
across many nodes, when not necessary, and this
was playing havoc with GPU test runs.

Set the default number of outstanding flux jobs
to be 8 per node rather than 1 per node.
This may be over-ridden by a project as needed
with the  --num_concurrent_jobs ats option.

Remove reference to numberNodesExclusivelyUsed,
as the -N option under flux does *not* grant
exclusive access to the node.  That is, flux
will schedule multiple jobs to run concurrently
on the same node, even if -N1 is specified.
Thus the concept of 'exclusive node use' is
not supported by the -N option and ATS will
not support it at this time.  We could do
so later if needed, as we can set the number
of cpus needed per rank as needed to get
exlusivity.