Track numNodesAvailable and numGPUsAvailable and user this in remainingCapacity to short keep down the call to the flux resource manager.
Also, if a user specified -nn directly, then
give the job exclusive access to the nodes.
That is add the --exclusive option to the flux
run line.  This concept of exclusive node usage
is a traditional ATS concept, and is honored
under LSF on Blueos as well, so this keeps
the meaning the same.

Do some re-arraning when setting
self.numberTestsRunningMax, self.npMax,
and self.npMaxH, which are used in the management.py
file.