Fix default setting of npMax with slurmProcessorScheduled
We should use the number of CPUs reported by slurm when
possible.  To do so, changed the lines as so:

  if self.runningWithinSalloc:
-  self.npMax = min(self.npMaxH, self.npMax)
+  self.npMax = max(self.npMaxH, self.npMax)

That is, take the max of npMaxHY and npMax, not the min.
self.npMax was set using a slurm env variable.

npMaxH is 36, which is the default if not in an
allocation (ie slurm not yet running) and if MACHINE_TYPE
was not specified.

This whould work better going forward, hopefully removing
the need to set MACHINE_TYPE for some projects.