Update opens used for readlines. (#167)
Update opens with errors='ignore' in attempt to stop the python
coding in ATS from dying when it reads output generated by applications
which are not utf-8 valid.  This can happen when applications print
garbage output -- but when this happens the python in ATS breaks
and the testing breaks.

Try to simply ignore such garbage output.