Add --cpx option.
For use on flux systems.  The --cpx option informs
ATS to use the appropriate publicly instgalled CPX mapfile
and to utilize 24 GPUs per node on ATS-4 systems.

The user must ensure the nodes to be used are allocated
with the CPX options:

--conf=resource.rediscover=true --setattr=gpumode=CPX

Then run using "flux --cpx <other options>"

In testing on rzadams, this increased throughput by more than
a facator of four.